requirements.txt
# requirements.txt
flask
flask-cors
requests
python-dotenv
lxml
beautifulsoup4
gunicorn